chore: move from example to integration
