+++ test/integration/PrivateKey.js
@@ -9,6 +9,8 @@ import {
     TransactionId,
     Mnemonic,
     AccountUpdateTransaction,
+    AccountCreateTransaction,
+    TransferTransaction,
 } from "../../src/exports.js";
 import dotenv from "dotenv";
 import IntegrationTestEnv from "./client/NodeIntegrationTestEnv.js";
@@ -120,5 +122,67 @@ describe("PrivateKey signTransaction", function () {
         ).getReceipt(env.client);
 
         expect(receipt.status).to.be.equal(Status.Success);
+    });
+
+    it("should create, sign, and execute a transaction with multiple signatures", async function () {
+        // Step 1: Generate private keys
+        const key1 = PrivateKey.generateED25519();
+        const key2 = PrivateKey.generateED25519();
+        const key3 = PrivateKey.generateED25519();
+        const key4 = PrivateKey.generateED25519();
+        const key5 = PrivateKey.generateED25519();
+        const recipient = PrivateKey.generateED25519().publicKey;
+
+        const keyList = new KeyList(
+            [
+                key1.publicKey,
+                key2.publicKey,
+                key3.publicKey,
+                key4.publicKey,
+                key5.publicKey,
+            ],
+            2,
+        );
+
+        // Step 2: Create account with key list
+        const { accountId } = await (
+            await new AccountCreateTransaction()
+                .setKeyWithoutAlias(keyList)
+                .setInitialBalance(new Hbar(1))
+                .freezeWith(env.client)
+                .execute(env.client)
+        ).getReceipt(env.client);
+
+        const { accountId: recipientId2 } = await (
+            await new AccountCreateTransaction()
+                .setKeyWithoutAlias(recipient)
+                .freezeWith(env.client)
+                .execute(env.client)
+        ).getReceipt(env.client);
+
+        // Step 3: Create and freeze transfer transaction
+        const transferTx = new TransferTransaction()
+            .addHbarTransfer(accountId, new Hbar(-1))
+            .addHbarTransfer(recipientId2, new Hbar(1))
+            .freezeWith(env.client);
+
+        // Step 4: Sign transaction
+        const signature1 = key1.signTransaction(transferTx, true);
+        const signature2 = key2.signTransaction(transferTx);
+
+        // Step 5: Add signatures and execute transaction
+        const { status } = await (
+            await transferTx
+                .addSignature(key1.publicKey, signature1)
+                .addSignature(key2.publicKey, signature2)
+                .execute(env.client)
+        ).getReceipt(env.client);
+
+        // Step 6: Verify transaction status
+        expect(status.toString()).to.be.equal("SUCCESS");
+    });
+
+    after(async function () {
+        await env.close();
     });
 });
